requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
python-multipart   # <--- add this line
